+++ packages/docs/app.config.ts
@@ -1,9 +1,9 @@
 export default defineAppConfig({
   docus: {
     title: "MustardCLI",
-    description: "IoC & Native ES Decotator based command-line app builder.",
+    description: "IoC & Native ES Decorator based command-line app builder.",
     url: "https://github.com/LinbuduLab/Mustard",
-    image: "/social-card-preview.png",
+    image: "/logo.png",
     socials: {
       github: "LinbuduLab",
     },
@@ -13,7 +13,7 @@ export default defineAppConfig({
       contributors: true,
     },
     header: {
-      title: "LinbuduLab Projects",
+      title: "LinbuduLab",
       logo: false,
       showLinkIcon: true,
     },