+++ packages/mustard-cli/source/Components/MustardUtilsProvider.ts
@@ -1,5 +1,6 @@
 import fs from "fs";
 import fsp from "fs/promises";
+import tty from "tty";
 import { EOL } from "os";
 
 import type { Nullable } from "../Typings/Shared.struct";
@@ -90,6 +91,112 @@ export class JSONHelper {
   }
 }
 
+// source: https://github.com/alexeyraspopov/picocolors/blob/main/picocolors.js
+export class ColorsHelper {
+  public static get isColorSupported(): boolean {
+    return (
+      !("NO_COLOR" in process.env || process.argv.includes("--no-color")) &&
+      ("FORCE_COLOR" in process.env ||
+        process.argv.includes("--color") ||
+        process.platform === "win32" ||
+        (tty.isatty(1) && process.env["TERM"] !== "dumb") ||
+        "CI" in process.env)
+    );
+  }
+
+  private static replaceClose = (
+    string: string,
+    close: string,
+    replace: string,
+    index: number
+  ): string => {
+    let start = string.substring(0, index) + replace;
+    let end = string.substring(index + close.length);
+    let nextIndex = end.indexOf(close);
+    return ~nextIndex
+      ? start + ColorsHelper.replaceClose(end, close, replace, nextIndex)
+      : start + end;
+  };
+
+  private static formatter =
+    (open: string, close: string, replace: string = open) =>
+    (input: string) => {
+      let string = "" + input;
+      let index = string.indexOf(close, open.length);
+      return ~index
+        ? open +
+            ColorsHelper.replaceClose(string, close, replace, index) +
+            close
+        : open + string + close;
+    };
+
+  private static factory = (open: string): ((input: string) => string) => {
+    return ColorsHelper.isColorSupported
+      ? ColorsHelper.formatter(open, "\x1b[39m")
+      : String;
+  };
+
+  private static bgFactory = (open: string): ((input: string) => string) => {
+    return ColorsHelper.isColorSupported
+      ? ColorsHelper.formatter(open, "\x1b[49m")
+      : String;
+  };
+
+  public static bold = ColorsHelper.isColorSupported
+    ? ColorsHelper.formatter("\x1b[1m", "\x1b[22m", "\x1b[22m\x1b[1m")
+    : String;
+  public static dim = ColorsHelper.isColorSupported
+    ? ColorsHelper.formatter("\x1b[2m", "\x1b[22m", "\x1b[22m\x1b[2m")
+    : String;
+  public static italic = ColorsHelper.isColorSupported
+    ? ColorsHelper.formatter("\x1b[3m", "\x1b[23m")
+    : String;
+  public static underline = ColorsHelper.isColorSupported
+    ? ColorsHelper.formatter("\x1b[4m", "\x1b[24m")
+    : String;
+  public static inverse = ColorsHelper.isColorSupported
+    ? ColorsHelper.formatter("\x1b[7m", "\x1b[27m")
+    : String;
+  public static hidden = ColorsHelper.isColorSupported
+    ? ColorsHelper.formatter("\x1b[8m", "\x1b[28m")
+    : String;
+  public static strikethrough = ColorsHelper.isColorSupported
+    ? ColorsHelper.formatter("\x1b[9m", "\x1b[29m")
+    : String;
+
+  public static black = ColorsHelper.factory("\x1b[30m");
+
+  public static red = ColorsHelper.factory("\x1b[31m");
+
+  public static green = ColorsHelper.factory("\x1b[32m");
+
+  public static yellow = ColorsHelper.factory("\x1b[33m");
+
+  public static blue = ColorsHelper.factory("\x1b[34m");
+
+  public static magenta = ColorsHelper.factory("\x1b[35m");
+
+  public static cyan = ColorsHelper.factory("\x1b[36m");
+
+  public static white = ColorsHelper.factory("\x1b[37m");
+
+  public static gray = ColorsHelper.factory("\x1b[90m");
+
+  public static bgRed = ColorsHelper.bgFactory("\x1b[41m");
+
+  public static bgGreen = ColorsHelper.bgFactory("\x1b[42m");
+
+  public static bgYellow = ColorsHelper.bgFactory("\x1b[43m");
+
+  public static bgBlue = ColorsHelper.bgFactory("\x1b[44m");
+
+  public static bgMagenta = ColorsHelper.bgFactory("\x1b[45m");
+
+  public static bgCyan = ColorsHelper.bgFactory("\x1b[46m");
+
+  public static bgWhite = ColorsHelper.bgFactory("\x1b[47m");
+}
+
 /**
  * @Utils
  * @Utils.Json
@@ -99,6 +206,7 @@ export class MustardUtilsProvider {
   public static produce() {
     return {
       json: MustardUtilsProvider.json,
+      colors: MustardUtilsProvider.colors,
     };
   }
 
@@ -109,5 +217,9 @@ export class MustardUtilsProvider {
       writeSync: JSONHelper.writeJsonSync,
       write: JSONHelper.writeJson,
     };
+  }
+
+  public static get colors() {
+    return ColorsHelper;
   }
 }