+++ packages/mustard-cli/source/Decorators/Input.ts
@@ -1,17 +1,45 @@
 import type { AnyClassFieldDecoratorReturnType } from "../Typings/Temp";
+import type { InputConfiguration } from "../Typings/Option.struct";
 
 export class InputDecorator {
+  /**
+   * TODO:
+   *
+   * @Input(name, alias, description)
+   * @Input({ name, alias, description })
+   *
+   * Inject inputs after commands
+   * @example
+   * class RunCommand {
+   *  \@Input('list of projects to include')
+   *   public projects: string[];
+   * }
+   */
+  public static Input(description?: string): AnyClassFieldDecoratorReturnType;
+
   /**
    * Inject inputs after commands
    * @example
    * class RunCommand {
-   *  \@Input()
-   *   public inputs: string[];
+   *  \@Input({ description: 'list of projects to include' } })
+   *   public projects: string[];
    * }
    */
-  public static Input(): AnyClassFieldDecoratorReturnType {
-    return (_, context) => () => ({
-      type: "Input",
-    });
+  public static Input(
+    configuration?: InputConfiguration
+  ): AnyClassFieldDecoratorReturnType;
+  public static Input(
+    config?: string | InputConfiguration
+  ): AnyClassFieldDecoratorReturnType {
+    const inputDescription =
+      typeof config === "string" ? config : config?.description;
+
+    return (_, context) => (initValue) => {
+      return {
+        type: "Input",
+        initValue,
+        description: inputDescription,
+      };
+    };
   }
 }