+++ packages/mustard-cli/source/Typings/Configuration.struct.ts
@@ -7,6 +7,8 @@ export interface Configurations {
   debug: boolean;
   enableUsage: boolean | ((registration?: CommandRegistryPayload) => string);
   enableVersion: false | MaybeFactory<string>;
+  ignoreValidationErrors: boolean;
+  defaultOverrides: boolean;
   lifeCycles?: Partial<LifeCycles>;
   didYouMean?: boolean;
   providers?: Provider[];
@@ -21,7 +23,7 @@ export interface LifeCycles {
 export interface CLIInstantiationConfiguration
   extends Partial<Configurations> {}
 
-export type CommandList = typeof CommandStruct[];
+export type CommandList = (typeof CommandStruct)[];
 
 export interface AppFactoryOptions {
   name?: string;