+++ packages/mustard-cli/source/__tests__/Fixtures/TestHelper.ts
@@ -1,0 +1,3 @@
+export class TestHelper {
+  public static IntegrationExecutor = "ts-node-esm --transpile-only";
+}