+++ packages/mustard-cli/source/__tests__/Integrations/Common/Common.spec.ts
@@ -1,5 +1,6 @@
 import { describe, it, expect } from "vitest";
 import { execaCommand } from "execa";
+import { TestHelper } from "../../Fixtures/TestHelper";
 import path from "path";
 
 const UsagePath = path.resolve(__dirname, "./Usage.ts");
@@ -7,7 +8,7 @@ const UsagePath = path.resolve(__dirname, "./Usage.ts");
 describe("IntegrationTesting:CommonCommandHandle", () => {
   it("should dispatch command", async () => {
     const { stdout: stdoutWithRoot1 } = await execaCommand(
-      `ts-node-esm ${UsagePath}`
+      `${TestHelper.IntegrationExecutor} ${UsagePath}`
     );
     expect(stdoutWithRoot1).toMatchInlineSnapshot(
       `
@@ -21,7 +22,7 @@ describe("IntegrationTesting:CommonCommandHandle", () => {
     );
 
     const { stdout: stdoutWithRoot2 } = await execaCommand(
-      `ts-node-esm ${UsagePath} input1 input2 input3`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} input1 input2 input3`
     );
     expect(stdoutWithRoot2).toMatchInlineSnapshot(
       `
@@ -35,7 +36,7 @@ describe("IntegrationTesting:CommonCommandHandle", () => {
     );
 
     const { stdout: stdoutWithRoot3 } = await execaCommand(
-      `ts-node-esm ${UsagePath} input1 input2 input3 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app4`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} input1 input2 input3 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app4`
     );
     expect(stdoutWithRoot3).toMatchInlineSnapshot(
       `
@@ -49,7 +50,7 @@ describe("IntegrationTesting:CommonCommandHandle", () => {
     );
 
     const { stdout: stdout1 } = await execaCommand(
-      `ts-node-esm ${UsagePath} run`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} run`
     );
     expect(stdout1).toMatchInlineSnapshot(
       `
@@ -63,7 +64,7 @@ describe("IntegrationTesting:CommonCommandHandle", () => {
     );
 
     const { stdout: stdout2 } = await execaCommand(
-      `ts-node-esm ${UsagePath} run input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app4`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} run input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app4`
     );
     expect(stdout2).toMatchInlineSnapshot(
       `
@@ -77,7 +78,7 @@ describe("IntegrationTesting:CommonCommandHandle", () => {
     );
 
     const { stdout: stdout3 } = await execaCommand(
-      `ts-node-esm ${UsagePath} update input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app4`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} update input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app4`
     );
     expect(stdout3).toMatchInlineSnapshot(
       `
@@ -90,7 +91,7 @@ describe("IntegrationTesting:CommonCommandHandle", () => {
     );
 
     const { stdout: stdout4 } = await execaCommand(
-      `ts-node-esm ${UsagePath} update input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app4`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} update input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app4`
     );
     expect(stdout4).toMatchInlineSnapshot(
       `
@@ -103,7 +104,7 @@ describe("IntegrationTesting:CommonCommandHandle", () => {
     );
 
     const { stdout: stdout5 } = await execaCommand(
-      `ts-node-esm ${UsagePath} update dep input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app4`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} update dep input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app4`
     );
     expect(stdout5).toMatchInlineSnapshot(
       `
@@ -117,7 +118,7 @@ describe("IntegrationTesting:CommonCommandHandle", () => {
     );
 
     const { stdout: stdout6 } = await execaCommand(
-      `ts-node-esm ${UsagePath} update dep node input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app4`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} update dep node input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app4`
     );
     expect(stdout6).toMatchInlineSnapshot(
       `
@@ -131,7 +132,7 @@ describe("IntegrationTesting:CommonCommandHandle", () => {
     );
 
     const { stdout: stdout7 } = await execaCommand(
-      `ts-node-esm ${UsagePath} update sys input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app4`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} update sys input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app4`
     );
     expect(stdout7).toMatchInlineSnapshot(
       `
@@ -145,7 +146,7 @@ describe("IntegrationTesting:CommonCommandHandle", () => {
     );
 
     const { stdout: stdout8 } = await execaCommand(
-      `ts-node-esm ${UsagePath} run`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} run`
     );
     expect(stdout8).toMatchInlineSnapshot(
       `
@@ -159,7 +160,7 @@ describe("IntegrationTesting:CommonCommandHandle", () => {
     );
 
     const { stdout: stdout9 } = await execaCommand(
-      `ts-node-esm ${UsagePath} r input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} r input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app`
     );
     expect(stdout9).toMatchInlineSnapshot(
       `
@@ -173,7 +174,7 @@ describe("IntegrationTesting:CommonCommandHandle", () => {
     );
 
     const { stdout: stdout10 } = await execaCommand(
-      `ts-node-esm ${UsagePath} u d input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app4`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} u d input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app4`
     );
     expect(stdout10).toMatchInlineSnapshot(
       `
@@ -187,7 +188,7 @@ describe("IntegrationTesting:CommonCommandHandle", () => {
     );
 
     const { stdout: stdout11 } = await execaCommand(
-      `ts-node-esm ${UsagePath} u d n input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app4`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} u d n input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app4`
     );
     expect(stdout11).toMatchInlineSnapshot(
       `