+++ packages/mustard-cli/source/__tests__/Integrations/NonCompleteParse/NonCompleteParse.spec.ts
@@ -1,5 +1,6 @@
 import { describe, it, expect } from "vitest";
 import { execaCommand } from "execa";
+import { TestHelper } from "../../Fixtures/TestHelper";
 import path from "path";
 
 const UsagePath = path.resolve(__dirname, "./Usage.ts");
@@ -7,7 +8,7 @@ const UsagePath = path.resolve(__dirname, "./Usage.ts");
 describe("IntegrationTesting:NonCompleteParse", () => {
   it("should dispatch command", async () => {
     const { stdout: stdoutWithRoot1 } = await execaCommand(
-      `ts-node-esm ${UsagePath}`
+      `${TestHelper.IntegrationExecutor} ${UsagePath}`
     );
     expect(stdoutWithRoot1).toMatchInlineSnapshot(
       `
@@ -20,7 +21,7 @@ describe("IntegrationTesting:NonCompleteParse", () => {
     );
 
     const { stdout: stdoutWithRoot2 } = await execaCommand(
-      `ts-node-esm ${UsagePath} input1 input2 input3`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} input1 input2 input3`
     );
     expect(stdoutWithRoot2).toMatchInlineSnapshot(
       `
@@ -33,7 +34,7 @@ describe("IntegrationTesting:NonCompleteParse", () => {
     );
 
     const { stdout: stdoutWithRoot3 } = await execaCommand(
-      `ts-node-esm ${UsagePath} input1 input2 input3 --msg Hello --pure pureValue`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} input1 input2 input3 --msg Hello --pure pureValue`
     );
     expect(stdoutWithRoot3).toMatchInlineSnapshot(
       `
@@ -46,7 +47,7 @@ describe("IntegrationTesting:NonCompleteParse", () => {
     );
 
     const { stdout: stdout1 } = await execaCommand(
-      `ts-node-esm ${UsagePath} run`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} run`
     );
     expect(stdout1).toMatchInlineSnapshot(
       `
@@ -59,7 +60,7 @@ describe("IntegrationTesting:NonCompleteParse", () => {
     );
 
     const { stdout: stdout2 } = await execaCommand(
-      `ts-node-esm ${UsagePath} run input1 input2 --msg Hello --pure pureValue`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} run input1 input2 --msg Hello --pure pureValue`
     );
     expect(stdout2).toMatchInlineSnapshot(
       `
@@ -72,7 +73,7 @@ describe("IntegrationTesting:NonCompleteParse", () => {
     );
 
     const { stdout: stdout3 } = await execaCommand(
-      `ts-node-esm ${UsagePath} update input1 input2 --msg Hello --pure pureValue`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} update input1 input2 --msg Hello --pure pureValue`
     );
     expect(stdout3).toMatchInlineSnapshot(
       `
@@ -84,7 +85,7 @@ describe("IntegrationTesting:NonCompleteParse", () => {
     );
 
     const { stdout: stdout4 } = await execaCommand(
-      `ts-node-esm ${UsagePath} update input1 input2 --msg Hello --pure pureValue`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} update input1 input2 --msg Hello --pure pureValue`
     );
     expect(stdout4).toMatchInlineSnapshot(
       `
@@ -96,7 +97,7 @@ describe("IntegrationTesting:NonCompleteParse", () => {
     );
 
     const { stdout: stdout5 } = await execaCommand(
-      `ts-node-esm ${UsagePath} update dep input1 input2 --msg Hello --pure pureValue`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} update dep input1 input2 --msg Hello --pure pureValue`
     );
     expect(stdout5).toMatchInlineSnapshot(
       `
@@ -109,7 +110,7 @@ describe("IntegrationTesting:NonCompleteParse", () => {
     );
 
     const { stdout: stdout6 } = await execaCommand(
-      `ts-node-esm ${UsagePath} update dep node input1 input2 --msg Hello --pure pureValue`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} update dep node input1 input2 --msg Hello --pure pureValue`
     );
     expect(stdout6).toMatchInlineSnapshot(
       `
@@ -122,7 +123,7 @@ describe("IntegrationTesting:NonCompleteParse", () => {
     );
 
     const { stdout: stdout7 } = await execaCommand(
-      `ts-node-esm ${UsagePath} update sys input1 input2 --msg Hello --pure pureValue`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} update sys input1 input2 --msg Hello --pure pureValue`
     );
     expect(stdout7).toMatchInlineSnapshot(
       `
@@ -135,7 +136,7 @@ describe("IntegrationTesting:NonCompleteParse", () => {
     );
 
     const { stdout: stdout8 } = await execaCommand(
-      `ts-node-esm ${UsagePath} run`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} run`
     );
     expect(stdout8).toMatchInlineSnapshot(
       `
@@ -148,7 +149,7 @@ describe("IntegrationTesting:NonCompleteParse", () => {
     );
 
     const { stdout: stdout9 } = await execaCommand(
-      `ts-node-esm ${UsagePath} r input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} r input1 input2 --msg Hello --pure pureValue --projects app1 app2 app3 --projects app`
     );
     expect(stdout9).toMatchInlineSnapshot(
       `
@@ -161,7 +162,7 @@ describe("IntegrationTesting:NonCompleteParse", () => {
     );
 
     const { stdout: stdout10 } = await execaCommand(
-      `ts-node-esm ${UsagePath} u d input1 input2 --msg Hello --pure pureValue`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} u d input1 input2 --msg Hello --pure pureValue`
     );
     expect(stdout10).toMatchInlineSnapshot(
       `
@@ -174,7 +175,7 @@ describe("IntegrationTesting:NonCompleteParse", () => {
     );
 
     const { stdout: stdout11 } = await execaCommand(
-      `ts-node-esm ${UsagePath} u d n input1 input2 --msg Hello --pure pureValue`
+      `${TestHelper.IntegrationExecutor} ${UsagePath} u d n input1 input2 --msg Hello --pure pureValue`
     );
     expect(stdout11).toMatchInlineSnapshot(
       `