+++ packages/mustard-cli/CHANGELOG.md
@@ -1,3 +1,58 @@
 # ChangeLog for mustard-cli
 
-## 1.0.0
+## 0.8.1
+
+### Patch Changes
+
+- Provide package-specified README
+
+## 0.8.0
+
+### Minor Changes
+
+- fix template dev workflow
+- enhance dev workflow
+
+## 0.7.1
+
+### Patch Changes
+
+- fixup package import path
+
+## 0.7.0
+
+### Minor Changes
+
+- 036f9b9: Fix starter dependencies
+
+## 0.6.3
+
+### Patch Changes
+
+- fixup \_\_dirname in es scope
+- 746ae49: Fix build config, use PURE ESM
+
+## 0.6.0
+
+### Minor Changes
+
+- use module:es2022
+
+## 0.5.0
+
+### Minor Changes
+
+- 4902a30: Fix required schema validation
+
+### Patch Changes
+
+- Bump typescript version to 5.0.0 beta
+- 7c32d55: Fix UsageGenerator for all suits
+- 7302020: Configuration and Usage info collection for @Input
+
+## 0.4.1
+
+### Patch Changes
+
+- b340fc2: add colors for built-in utils
+- c469b63: Testing